Added two folders
Added folder BPA end caps. This is taken from the AARL puppy model. Parts 18_02_01_010 and _011 are the port and cap, respectively, for the Festo 10 mm BPAs. Similarly, parts _012 and _013 are for the Festo 20 mm BPAs, and parts _014 and _015 are for the Festo 40 mm BPAs. Use the ones that say U10 or M5 (push to connect adapters) for 012-015, don't use the ones that say NPT.

Added folder Living Machines 2021 - BPA Test for the components that are going to be used in the BPA characterization static test being headed by Lawrence.