Organized static characterization
Further organized static characterization folder. Created folder to place all solid models in. Created a folder for code, with sub-folders for arduino or matlab. Code folder mainly has ValvePWM and serial communication files. Data processing code is contained in other locations. It's not clear which data is good and which needs to be deleted.